+++ lib/rabarber/access.rb
@@ -2,25 +2,25 @@
 
 module Rabarber
   module Access
-    def access_granted?(roles, controller, action, custom_rule_receiver)
-      controller_accessible?(roles, controller, custom_rule_receiver) ||
-        action_accessible?(roles, controller, action, custom_rule_receiver)
+    def access_granted?(roles, controller, action, dynamic_rule_receiver)
+      controller_accessible?(roles, controller, dynamic_rule_receiver) ||
+        action_accessible?(roles, controller, action, dynamic_rule_receiver)
     end
 
-    def controller_accessible?(roles, controller, custom_rule_receiver)
-      accessible_controllers(roles, custom_rule_receiver).any? do |accessible_controller|
+    def controller_accessible?(roles, controller, dynamic_rule_receiver)
+      accessible_controllers(roles, dynamic_rule_receiver).any? do |accessible_controller|
         controller <= accessible_controller
       end
     end
 
-    def action_accessible?(roles, controller, action, custom_rule_receiver)
-      action_rules[controller].any? { |rule| rule.verify_access(roles, custom_rule_receiver, action) }
+    def action_accessible?(roles, controller, action, dynamic_rule_receiver)
+      action_rules[controller].any? { |rule| rule.verify_access(roles, dynamic_rule_receiver, action) }
     end
 
     private
 
-    def accessible_controllers(roles, custom_rule_receiver)
-      controller_rules.select { |_, rule| rule.verify_access(roles, custom_rule_receiver) }.keys
+    def accessible_controllers(roles, dynamic_rule_receiver)
+      controller_rules.select { |_, rule| rule.verify_access(roles, dynamic_rule_receiver) }.keys
     end
   end
 end