+++ spec/rabarber/controllers/concerns/authorization_spec.rb
@@ -3,6 +3,57 @@
 RSpec.describe Rabarber::Authorization do
   let(:user) { User.create! }
 
+  describe ".grant_access" do
+    subject { DummyController.grant_access(**args) }
+
+    context "when arguments are valid" do
+      context "when 'if' is specified" do
+        let(:args) { { action: :foo, roles: :bar, if: -> { true } } }
+
+        it "writes the permission" do
+          expect(::Rabarber::Permissions).to receive(:write).with(DummyController, :foo, :bar, args[:if], nil)
+          subject
+        end
+      end
+
+      context "when 'unless' is specified" do
+        let(:args) { { action: :foo, roles: :bar, unless: -> { false } } }
+
+        it "writes the permission" do
+          expect(::Rabarber::Permissions).to receive(:write).with(DummyController, :foo, :bar, nil, args[:unless])
+          subject
+        end
+      end
+
+      context "when neither 'if' nor 'unless' is specified" do
+        let(:args) { { action: :foo, roles: :bar } }
+
+        it "writes the permission" do
+          expect(::Rabarber::Permissions).to receive(:write).with(DummyController, :foo, :bar, nil, nil)
+          subject
+        end
+      end
+
+      context "when both 'if' and 'unless' are specified" do
+        let(:args) { { action: :foo, roles: :bar, if: -> { true }, unless: -> { false } } }
+
+        it "writes the permission" do
+          expect(::Rabarber::Permissions).to receive(:write).with(DummyController, :foo, :bar, args[:if], args[:unless])
+          subject
+        end
+      end
+
+      context "when action and roles are omitted" do
+        let(:args) { {} }
+
+        it "writes the permission" do
+          expect(::Rabarber::Permissions).to receive(:write).with(DummyController, nil, nil, nil, nil)
+          subject
+        end
+      end
+    end
+  end
+
   shared_examples_for "it allows access" do |hash|
     it "allows access when request format is html" do
       send(hash.keys.first, hash.values.first, params: hash[:params])
@@ -104,36 +155,74 @@ RSpec.describe Rabarber::Authorization do
       it_behaves_like "it does not allow access when user must have roles", delete: :no_access
     end
 
-    describe "when a custom rule is defined as a lambda" do
-      context "when the lambda returns true" do
-        before { user.assign_roles(:admin) }
+    context "when dynamic rule is not negated" do
+      describe "when dynamic rule is defined as a lambda" do
+        context "when the lambda returns true" do
+          before { user.assign_roles(:admin) }
 
-        it_behaves_like "it allows access", get: :if_lambda, params: { foo: "bar" }
+          it_behaves_like "it allows access", get: :if_lambda, params: { foo: "bar" }
+        end
+
+        context "when the lambda returns false" do
+          before { user.assign_roles(:admin) }
+
+          it_behaves_like "it does not allow access", get: :if_lambda, params: { foo: "baz" }
+        end
+
+        it_behaves_like "it does not allow access when user must have roles", get: :if_lambda, params: { foo: "bar" }
       end
 
-      context "when the lambda returns false" do
-        before { user.assign_roles(:admin) }
+      describe "when dynamic rule is defined as a method" do
+        context "when the method returns true" do
+          before { user.assign_roles(:admin) }
 
-        it_behaves_like "it does not allow access", get: :if_lambda, params: { foo: "baz" }
+          it_behaves_like "it allows access", post: :if_method, params: { bad: "baz" }
+        end
+
+        context "when the method returns false" do
+          before { user.assign_roles(:admin) }
+
+          it_behaves_like "it does not allow access", post: :if_method, params: { bad: "bar" }
+        end
+
+        it_behaves_like "it does not allow access when user must have roles", post: :if_method, params: { bad: "baz" }
       end
-
-      it_behaves_like "it does not allow access when user must have roles", get: :if_lambda, params: { foo: "bar" }
     end
 
-    describe "when a custom rule is defined as a method" do
-      context "when the method returns true" do
-        before { user.assign_roles(:admin) }
+    context "when dynamic rule is negated" do
+      describe "when dynamic rule is defined as a lambda" do
+        context "when the lambda returns true" do
+          before { user.assign_roles(:admin) }
 
-        it_behaves_like "it allows access", post: :if_method, params: { bad: "baz" }
+          it_behaves_like "it does not allow access", patch: :unless_lambda, params: { foo: "bar" }
+        end
+
+        context "when the lambda returns false" do
+          before { user.assign_roles(:admin) }
+
+          it_behaves_like "it allows access", patch: :unless_lambda, params: { foo: "baz" }
+        end
+
+        it_behaves_like "it does not allow access when user must have roles",
+                        patch: :unless_lambda, params: { foo: "bar" }
       end
 
-      context "when the method returns false" do
-        before { user.assign_roles(:admin) }
+      describe "when dynamic rule is defined as a method" do
+        context "when the method returns true" do
+          before { user.assign_roles(:admin) }
 
-        it_behaves_like "it does not allow access", post: :if_method, params: { bad: "bar" }
+          it_behaves_like "it does not allow access", delete: :unless_method, params: { bad: "baz" }
+        end
+
+        context "when the method returns false" do
+          before { user.assign_roles(:admin) }
+
+          it_behaves_like "it allows access", delete: :unless_method, params: { bad: "bar" }
+        end
+
+        it_behaves_like "it does not allow access when user must have roles",
+                        delete: :unless_method, params: { bad: "baz" }
       end
-
-      it_behaves_like "it does not allow access when user must have roles", post: :if_method, params: { bad: "baz" }
     end
   end