+++ spec/rabarber/rule_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe Rabarber::Rule do
     context "when action is invalid" do
       [1, ["index"], "", Symbol, {}, :""].each do |wrong_action_name|
         it "raises an error when '#{wrong_action_name}' is given as an action name" do
-          expect { described_class.new(wrong_action_name, nil, nil) }.to raise_error(
+          expect { described_class.new(wrong_action_name, nil, nil, nil) }.to raise_error(
             Rabarber::InvalidArgumentError, "Action name must be a Symbol or a String"
           )
         end
@@ -15,7 +15,7 @@ RSpec.describe Rabarber::Rule do
     context "when roles are invalid" do
       [1, Symbol, "Admin", :"foo-bar", "", [""], ["admin "]].each do |wrong_roles|
         it "raises an error when '#{wrong_roles}' are given as roles" do
-          expect { described_class.new(nil, wrong_roles, nil) }.to raise_error(
+          expect { described_class.new(nil, wrong_roles, nil, nil) }.to raise_error(
             Rabarber::InvalidArgumentError,
             "Role names must be Symbols or Strings and may only contain lowercase letters, numbers and underscores"
           )
@@ -23,12 +23,23 @@ RSpec.describe Rabarber::Rule do
       end
     end
 
-    context "when custom rule is invalid" do
-      [1, ["rule"], "", :"", Symbol, [], {}].each do |wrong_custom_rule|
-        it "raises an error when '#{wrong_custom_rule}' is given as a custom rule" do
-          expect { described_class.new(nil, nil, wrong_custom_rule) }.to raise_error(
+    context "when dynamic rule is invalid" do
+      [1, ["rule"], "", :"", Symbol, [], {}].each do |wrong_dynamic_rule|
+        it "raises an error when '#{wrong_dynamic_rule}' is given as a dynamic rule" do
+          expect { described_class.new(nil, nil, wrong_dynamic_rule, nil) }.to raise_error(
             Rabarber::InvalidArgumentError,
-            "Custom rule must be a Symbol, a String, or a Proc"
+            "Dynamic rule must be a Symbol, a String, or a Proc"
+          )
+        end
+      end
+    end
+
+    context "when negated dynamic rule is invalid" do
+      [1, ["rule"], "", :"", Symbol, [], {}].each do |wrong_dynamic_rule|
+        it "raises an error when '#{wrong_dynamic_rule}' is given as a dynamic rule" do
+          expect { described_class.new(nil, nil, nil, wrong_dynamic_rule) }.to raise_error(
+            Rabarber::InvalidArgumentError,
+            "Dynamic rule must be a Symbol, a String, or a Proc"
           )
         end
       end
@@ -38,13 +49,13 @@ RSpec.describe Rabarber::Rule do
   describe "#verify_access" do
     subject { rule.verify_access(:admin, DummyController, :index) }
 
-    let(:rule) { described_class.new(:index, :admin, -> { true }) }
+    let(:rule) { described_class.new(:index, :admin, -> { true }, nil) }
 
     context "if all conditions are met" do
       before do
         allow(rule).to receive(:action_accessible?).with(:index).and_return(true)
         allow(rule).to receive(:roles_permitted?).with(:admin).and_return(true)
-        allow(rule).to receive(:custom_rule_followed?).with(DummyController).and_return(true)
+        allow(rule).to receive(:dynamic_rule_followed?).with(DummyController).and_return(true)
       end
 
       it "returns true" do
@@ -56,8 +67,8 @@ RSpec.describe Rabarber::Rule do
       context "if action is not accessible" do
         before do
           allow(rule).to receive(:action_accessible?).with(:index).and_return(false)
-          allow(rule).to receive(:roles_permitted?).with(:admin).and_return([true, false].sample)
-          allow(rule).to receive(:custom_rule_followed?).with(DummyController).and_return([true, false].sample)
+          allow(rule).to receive(:roles_permitted?).with(:admin).and_return(true)
+          allow(rule).to receive(:dynamic_rule_followed?).with(DummyController).and_return(true)
         end
 
         it "returns false" do
@@ -67,9 +78,9 @@ RSpec.describe Rabarber::Rule do
 
       context "if roles are not permitted" do
         before do
-          allow(rule).to receive(:action_accessible?).with(:index).and_return([true, false].sample)
+          allow(rule).to receive(:action_accessible?).with(:index).and_return(true)
           allow(rule).to receive(:roles_permitted?).with(:admin).and_return(false)
-          allow(rule).to receive(:custom_rule_followed?).with(DummyController).and_return([true, false].sample)
+          allow(rule).to receive(:dynamic_rule_followed?).with(DummyController).and_return(true)
         end
 
         it "returns false" do
@@ -77,11 +88,11 @@ RSpec.describe Rabarber::Rule do
         end
       end
 
-      context "if custom rule is not followed" do
+      context "if dynamic rule is not followed" do
         before do
-          allow(rule).to receive(:action_accessible?).with(:index).and_return([true, false].sample)
-          allow(rule).to receive(:roles_permitted?).with(:admin).and_return([true, false].sample)
-          allow(rule).to receive(:custom_rule_followed?).with(DummyController).and_return(false)
+          allow(rule).to receive(:action_accessible?).with(:index).and_return(true)
+          allow(rule).to receive(:roles_permitted?).with(:admin).and_return(true)
+          allow(rule).to receive(:dynamic_rule_followed?).with(DummyController).and_return(false)
         end
 
         it "returns false" do
@@ -94,7 +105,7 @@ RSpec.describe Rabarber::Rule do
   describe "#action_accessible?" do
     subject { rule.action_accessible?(action_name) }
 
-    let(:rule) { described_class.new(:index, [:admin], -> { true }) }
+    let(:rule) { described_class.new(:index, [:admin], -> { true }, nil) }
 
     context "if action is accesible" do
       let(:action_name) { :index }
@@ -124,7 +135,7 @@ RSpec.describe Rabarber::Rule do
   describe "#roles_permitted?" do
     subject { rule.roles_permitted?(user_roles) }
 
-    let(:rule) { described_class.new(:index, roles, nil) }
+    let(:rule) { described_class.new(:index, roles, nil, nil) }
 
     context "if roles are permitted" do
       let(:roles) { :admin }
@@ -187,59 +198,160 @@ RSpec.describe Rabarber::Rule do
     end
   end
 
-  describe "#custom_rule_followed?" do
-    subject { rule.custom_rule_followed?(custom_rule_receiver) }
+  describe "#dynamic_rule_followed?" do
+    subject { rule.dynamic_rule_followed?(dynamic_rule_receiver) }
 
-    let(:custom_rule_receiver) { double }
-    let(:rule) { described_class.new(:index, :manager, custom_rule) }
+    let(:dynamic_rule_receiver) { double }
+    let(:rule) { described_class.new(:index, :manager, dynamic_rule, negated_dynamic_rule) }
 
-    context "custom rule is empty" do
-      let(:custom_rule) { nil }
+    context "both dynamic rules are empty" do
+      let(:dynamic_rule) { nil }
+      let(:negated_dynamic_rule) { nil }
 
       it "returns true" do
         expect(subject).to be true
       end
     end
 
-    context "when custom rule is a proc" do
-      before { allow(custom_rule_receiver).to receive(:params).and_return({ foo: "bar" }) }
+    context "when dynamic rule is a proc" do
+      before { allow(dynamic_rule_receiver).to receive(:params).and_return({ foo: "bar" }) }
 
-      context "custom rule is followed" do
-        let(:custom_rule) { -> { params[:foo] == "bar" } }
+      context "when negated dynamic rule is nil" do
+        let(:negated_dynamic_rule) { nil }
 
-        it "returns true" do
-          expect(subject).to be true
+        context "when dynamic rule is nil" do
+          let(:dynamic_rule) { nil }
+
+          it { is_expected.to be true }
+        end
+
+        context "when dynamic rule is followed" do
+          let(:dynamic_rule) { -> { params[:foo] == "bar" } }
+
+          it { is_expected.to be true }
+        end
+
+        context "when dynamic rule is not followed" do
+          let(:dynamic_rule) { -> { params[:foo] == "baz" } }
+
+          it { is_expected.to be false }
         end
       end
 
-      context "custom rule is not followed" do
-        let(:custom_rule) { -> { params[:foo] == "bad" } }
+      context "when negated dynamic rule is followed" do
+        let(:negated_dynamic_rule) { -> { params[:foo] == "baz" } }
 
-        it "returns false" do
-          expect(subject).to be false
+        context "when dynamic rule is nil" do
+          let(:dynamic_rule) { nil }
+
+          it { is_expected.to be true }
+        end
+
+        context "when dynamic rule is followed" do
+          let(:dynamic_rule) { -> { params[:foo] == "bar" } }
+
+          it { is_expected.to be true }
+        end
+
+        context "when dynamic rule is not followed" do
+          let(:dynamic_rule) { -> { params[:foo] == "baz" } }
+
+          it { is_expected.to be false }
+        end
+      end
+
+      context "when negated dynamic rule is not followed" do
+        let(:negated_dynamic_rule) { -> { params[:foo] == "bar" } }
+
+        context "when dynamic rule is nil" do
+          let(:dynamic_rule) { nil }
+
+          it { is_expected.to be false }
+        end
+
+        context "when dynamic rule is followed" do
+          let(:dynamic_rule) { -> { params[:foo] == "bar" } }
+
+          it { is_expected.to be false }
+        end
+
+        context "when dynamic rule is not followed" do
+          let(:dynamic_rule) { -> { params[:foo] == "baz" } }
+
+          it { is_expected.to be false }
         end
       end
     end
 
-    context "when custom rule is a method name" do
+    context "when dynamic rule is a method name" do
       before do
-        allow(custom_rule_receiver).to receive(:foo).and_return(true)
-        allow(custom_rule_receiver).to receive(:bar).and_return(false)
+        allow(dynamic_rule_receiver).to receive(:foo).and_return(true)
+        allow(dynamic_rule_receiver).to receive(:bar).and_return(false)
       end
 
-      context "custom rule is followed" do
-        let(:custom_rule) { :foo }
+      context "when negated dynamic rule is nil" do
+        let(:negated_dynamic_rule) { nil }
 
-        it "returns true" do
-          expect(subject).to be true
+        context "when dynamic rule is nil" do
+          let(:dynamic_rule) { nil }
+
+          it { is_expected.to be true }
+        end
+
+        context "when dynamic rule is followed" do
+          let(:dynamic_rule) { :foo }
+
+          it { is_expected.to be true }
+        end
+
+        context "when dynamic rule is not followed" do
+          let(:dynamic_rule) { :bar }
+
+          it { is_expected.to be false }
         end
       end
 
-      context "custom rule is not followed" do
-        let(:custom_rule) { :bar }
+      context "when negated dynamic rule is followed" do
+        let(:negated_dynamic_rule) { :bar }
 
-        it "returns false" do
-          expect(subject).to be false
+        context "when dynamic rule is nil" do
+          let(:dynamic_rule) { nil }
+
+          it { is_expected.to be true }
+        end
+
+        context "when dynamic rule is followed" do
+          let(:dynamic_rule) { :foo }
+
+          it { is_expected.to be true }
+        end
+
+        context "when dynamic rule is not followed" do
+          let(:dynamic_rule) { :bar }
+
+          it { is_expected.to be false }
+        end
+      end
+
+      context "when negated dynamic rule is not followed" do
+        let(:negated_dynamic_rule) { :foo }
+
+        context "when dynamic rule is nil" do
+          let(:dynamic_rule) { nil }
+
+          it { is_expected.to be false }
+        end
+
+        context "when dynamic rule is followed" do
+          let(:dynamic_rule) { :foo }
+
+          it { is_expected.to be false }
+        end
+
+        context "when dynamic rule is not followed" do
+          let(:dynamic_rule) { :bar }
+
+          it { is_expected.to be false }
         end
       end
     end