+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+- Add support for `unless` argument in `grant_access` method, allowing to define negated dynamic rules
+- Fix a bug where specifying a dynamic rule as a symbol without specifying an action would result in an error
+
 ## 1.0.5
 
 - Add co-author: [trafium](https://github.com/trafium)